Fixed issue #20
Basically, when you're comparing 2 operands in the IF, and one of them is a
BOOL, fall back to the boolean comparison, and not the double comparison.

PHP supports type juggling, so does blitz. If you compare 3.0 with the string
"3", this results in a true, as well if you compare 0 with the string "0".
To enable this, we cast the string "3" to a double, with the atof system call.
It's very similar to intval() in PHP. But, when your string doesn't start with
a number, it returns double value 0.0.

So, when you have the string "house", it's double value is 0.0. If you compare
it with a boolean operator TRUE, you'll get a mismatch since 0.0 != 1.0.

So, whenever we encounter a BOOL operator now, we use the boolean value. For a
string this is checking if it's length is non-zero.